finish final adjustments to the base search pages
