+++ src/components/ui/home/search_container/SearchBox.scss
@@ -2,7 +2,7 @@
     //border: $border-red-solid;
     @include box-reset-default;
     height: 100%;
-    padding: 0px 15px;
+    padding: 0px 15px 15px 15px;
     #searchbar-row {
         //border: $border-yellow-solid;
         @include box-reset-default;
