+++ src/components/ui/home/search_container/general_results/GeneralResultsPage.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import IMG from "../../../../../assets/images/ImagesHUB";
-import Artists from "./list_components/Artists";
-import Albuns from "./list_components/Albuns";
-import Songs from "./list_components/Songs";
+import ArtistResults from "./list_components/ArtistResults";
+import AlbumResults from "./list_components/AlbumResults";
+import SongResults from "./list_components/SongResults";
 
 function GeneralResultsPage() {
     return (
@@ -10,20 +10,17 @@ function GeneralResultsPage() {
             <div id="gen-results-container" className="container-fluid">
                 <div id="gen-results-artists" className="row">
                     <div id="gen-results-artists-col" className="col">
-                        <h4>artists:</h4>
-                        <Artists/>
+                        <ArtistResults/>
                     </div>
                 </div>
                 <div id="gen-results-albuns" className="row">
                     <div id="gen-results-albuns-col" className="col">
-                        <h4>albuns:</h4>
-                        <Albuns/>
+                        <AlbumResults/>
                     </div>
                 </div>
                 <div id="gen-results-songs" className="row">
                     <div id="gen-results-songs-col" className="col">
-                        <h4>songs:</h4>
-                        <Songs/>
+                        <SongResults/>
                     </div>
                 </div>
             </div>
