+++ src/components/ui/home/search_container/general_results/GeneralResultsPage.scss
@@ -14,23 +14,6 @@
                 border-bottom: 1px solid white;
                 height: 30px;
             }
-            #artist-inner-row {
-                //border: $border-yellow-dotted;
-                @include box-reset-default;
-                height: 75px;
-                #artist-thumbnail {
-                    //border: $border-cyan-dotted;
-                    @include box-reset-default;
-                }
-                #artist-name {
-                    //border: $border-cyan-dotted;
-                    @include box-reset-default;
-                }
-            }
-            #artist-inner-row:hover {
-                background-color: $bg-transparent;
-                border-radius: 15px;
-            }
         }
     }  
     #gen-results-albuns {
@@ -44,23 +27,6 @@
                 @include box-reset-default;
                 border-bottom: 1px solid white;
                 height: 30px;
-            }
-            #albuns-inner-row {
-                //border: $border-yellow-dotted;
-                @include box-reset-default;
-                height: 75px;
-                #album-thumbnail {
-                    //border: $border-cyan-dotted;
-                    @include box-reset-default;
-                }
-                #album-title {
-                    //border: $border-cyan-dotted;
-                    @include box-reset-default;
-                }
-            }
-            #albuns-inner-row:hover {
-                background-color: $bg-transparent;
-                border-radius: 15px;
             }
         }
     }
@@ -77,123 +43,7 @@
                 margin-bottom: 5px;
                 height: 30px;
             }
-            #songs-inner-row {
-                //border: $border-yellow-dashed;
-                @include box-reset-default;
-                height: 50px;
-                #col-add {
-                    //border: $border-lime-dotted;
-                    @include box-reset-default;
-                    width: 40px;
-                    position: relative;
-                    #plus-icon {
-                        //border: $border-lime-dotted;
-                        @include box-reset-default;
-                        position: absolute;
-                    }
-                    #drag-icon {
-                        //border: $border-lime-dotted;
-                        @include box-reset-default;
-                        visibility: hidden;
-                        position: absolute;
-                    }
-                }
-                #col-cover {
-                    //border: $border-lime-dotted;
-                    @include box-reset-default;
-                    width: 50px;
-                    img {
-                        //border: $border-lime-dotted;
-                        @include box-reset-default;
-                    }
-                }
-                #col-title {
-                    //border: $border-lime-dotted;
-                    @include box-reset-default;
-                    white-space: nowrap;
-                    text-overflow: ellipsis;
-                    overflow: hidden;
-                    flex-wrap: wrap;
-                    h5 {
-                        //border: $border-magenta-dotted;
-                        @include box-reset-default;
-                        text-overflow: ellipsis;
-                        overflow: hidden;
-                        width: 100%;
-                        margin-bottom: -3px;
-                    }
-                    p {
-                        //border: $border-magenta-dotted;
-                        @include box-reset-default;
-                        text-overflow: ellipsis;
-                        overflow: hidden;
-                        width: 100%;
-                        margin-top: -3px;
-                    }
-                }
-                #col-album {
-                    //border: $border-lime-dotted;
-                    @include box-reset-default;
-                    white-space: nowrap;
-                    p {
-                        //border: $border-lime-dotted;
-                        @include box-reset-default;
-                        text-overflow: ellipsis;
-                        overflow: hidden;
-                    }
-                }
-                #col-duration {
-                    //border: $border-lime-dotted;
-                    @include box-reset-default;
-                    white-space: nowrap;
-                    width: 55px;
-                    p {
-                        //border: $border-lime-dotted;
-                        @include box-reset-default;
-                        text-overflow: ellipsis;
-                        overflow: hidden;
-                    }
-                }
-                #col-options {
-                    //border: $border-lime-dotted;
-                    @include box-reset-default;
-                    white-space: nowrap;
-                    width: 25px;
-                    margin-right: 5px;
-                    position: relative;
-                    #opt-icon {
-                        //border: $border-lime-dotted;
-                        @include box-reset-default;
-                    }
-                }
-                #col-add, #col-cover, #col-title, 
-                #col-album, #col-options, #col-duration {
-                    height: 100%;
-                    padding: 0 5px;
-                };  
-            }
-            #songs-inner-row:hover {
-                background-color: $bg-transparent;
-                border-radius: 15px;
-                #col-add {
-                    //border: $border-lime-dotted;
-                    @include box-reset-default;
-                    width: 40px;
-                    position: relative;
-                    #plus-icon {
-                        //border: $border-lime-dotted;
-                        @include box-reset-default;
-                        visibility: hidden;
-                        position: absolute;
-                    }
-                    #drag-icon {
-                        //border: $border-lime-dotted;
-                        @include box-reset-default;
-                        visibility: visible;
-                        position: absolute;
-                    }
-                }
-            }
+            
         }
     }
 }
