+++ src/components/ui/home/search_container/general_results/list_components/AlbumResults.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import IMG from "../../../../../../assets/images/ImagesHUB";
+import Albuns from "./unit_components/Albuns";
+
+function AlbumResults() {
+    return (
+        <>
+            <h4>albuns:</h4>
+            <Albuns/>
+        </>
+    );
+}
+
+export default AlbumResults;
