--- src/components/ui/home/search_container/general_results/list_components/Albuns.jsx
@@ -1,19 +1,0 @@
-import React from "react";
-import IMG from "../../../../../../assets/images/ImagesHUB";
-
-function Albuns() {
-    return (
-        <>
-            <div id="albuns-inner-row" className="row">
-                <div id="album-thumbnail" className="col">
-                
-                </div>
-                <div id="album-title" className="col">
-
-                </div>
-            </div>  
-        </>
-    );
-}
-
-export default Albuns;
