+++ src/components/ui/home/search_container/general_results/list_components/ArtistResults.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import IMG from "../../../../../../assets/images/ImagesHUB";
+import Artists from "./unit_components/Artists";
+
+function ArtistResults() {
+    return (
+        <>
+            <h4>artists:</h4>
+            <Artists/>
+        </>
+    );
+}
+
+export default ArtistResults;
