--- src/components/ui/home/search_container/general_results/list_components/Artists.jsx
@@ -1,19 +1,0 @@
-import React from "react";
-import IMG from "../../../../../../assets/images/ImagesHUB";
-
-function Artists() {
-    return (
-        <>
-            <div id="artist-inner-row" className="row">
-                <div id="artist-thumbnail" className="col">
-
-                </div>
-                <div id="artist-name" className="col">
-
-                </div>
-            </div>
-        </>
-    );
-}
-
-export default Artists;
