+++ src/components/ui/home/search_container/general_results/list_components/SongResults.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import IMG from "../../../../../../assets/images/ImagesHUB";
+import Songs from "./unit_components/Songs";
+
+function SongResults() {
+    return (
+        <>
+            <h4>songs:</h4>
+            <Songs/>
+        </>
+    );
+}
+
+export default SongResults;
