+++ src/components/ui/home/search_container/non_results/EmptyResultsPage.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import IMG from "../../../../assets/images/ImagesHUB";
+import IMG from "../../../../../assets/images/ImagesHUB";
 
 function EmptyResultsPage() {
     return (
