+++ src/scss/main.scss
@@ -25,6 +25,7 @@ $form-check-input-focus-box-shadow: none;
 @import "../components/ui/home/track_player/TrackPlayer.scss";
 @import "../components//ui/home/search_container/SearchBox.scss";
 @import "../components//ui/home/search_container/general_results/GeneralResultsPage.scss";
+@import "../components//ui/home/search_container/general_results/list_components/ListComponents.scss";
 
 //-----------login_page-----------//
